# services/backend/MathLLMBackend.Presentation_openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "MathLLMBackend.Presentation | openapi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Chat/create": {
      "post": {
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDto"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/Chat/get": {
      "get": {
        "tags": [
          "Chat"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/Chat/get/{chatId}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDto"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/Chat/{id}": {
      "delete": {
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          { }
        ]
      }
    },
...
    "/api/Stats/task-mode-titles": {
      "get": {
        "tags": [
          "Stats"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/Stats/user-stats": {
      "get": {
        "tags": [
          "Stats"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/Stats/user-details/{userId}": {
      "get": {
        "tags": [
          "Stats"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          { }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatDto": {
        "required": [
          "id",
          "name",
          "type",
          "taskType",
          "theoryLink"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "taskType": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskType"
              }
            ]
          },
          "theoryLink": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CreateChatRequestDto": {
        "required": [
          "name",
          "problemId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "problemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "TaskType": {
        "enum": [
          "Default",
          "Learning",
          "Guided",
          "Exam"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer [space] {your token}'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}